Refactor Goals Files, Link Dev API URI
